make functions have single responsibility
