+++ index.html
@@ -17,7 +17,7 @@
   <nav>
     <div class="nav-wrapper cyan darken-4">
       <a href="#" class="brand-logo left"> &nbsp; Inverted Index</a>
-      <a href="#modal1" class="right">Instructions</a>
+      <a href="#modal1" class="right padInstr">Instructions</a>
       <i class="material-icons right">info_outline</i>
     </div>
   </nav>
@@ -41,7 +41,7 @@
             <input type="file" id="files_id" accept=".json">
           </div>
           <div class="file-path-wrapper">
-            <input class="file-path validate" type="text" id="filename_id" placeholder="Upload one or more files"
+            <input class="file-path validate" type="text" id="filename_id" placeholder="Upload files here"
               value="">
           </div>
         </div>
@@ -110,9 +110,10 @@
   <script type="text/javascript" src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
   <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.8/js/materialize.min.js"></script>
   <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/angular-materialize/0.2.2/angular-materialize.min.js"></script>
-  <script type="text/javascript" src="src/inverted-index.js"></script>
-  <script type="text/javascript" src="src/app.js"></script>
   <script type="text/javascript" src="src/eventListener.js"></script>
+  <script type="text/javascript" src="src/util.js"></script>
+  <script type="text/javascript" src="src/invertedIndexUI.js"></script>
+  <script type="text/javascript" src="src/inverted-index.js"></script>
 </body>
 
 </html>
