+++ jasmine/SpecRunner.html
@@ -11,8 +11,11 @@
   <script src="lib/jasmine-2.3.4/boot.js"></script>
   <!-- include source files here... -->
   <script src="../src/inverted-index.js"></script>
-  <!-- include spec files here... -->
+  <script src="../src/app.js"></script>
+  <script src="../src/eventListener.js"></script>
+   <!-- include spec files here... -->
   <script src="build/bundle.js"></script>
+  <!--<script src="classbuild/bundle.js"></script>-->
 </head>
 
 <body>
